REFACTOR data
