+++ src/components/event.js
@@ -1,5 +1,6 @@
 import {TimeValue} from '../utils/utils.js';
 import {eventsData} from '../eventsList.js';
+import {eventsData} from '../data.js';
 import AbstractComponent from './abstract.js';
