+++ src/components/trip-info.js
@@ -1,3 +1,5 @@
+import {eventsData} from '../data.js';
+// import {createElement} from '../utils/utils.js';
 import AbstractComponent from './abstract.js';
